[Feat]: created new thunk for fetched data
- Created new thunk for fetched data from the firebase.
- This thunk will returns the updated cart and replaced it to the main cart.
