+++ redux-cart-app/src/store/cart-slice.js
@@ -1,5 +1,4 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { uiActions } from "./ui-slice";
 
 const cartSlice= createSlice({
   name: 'cart',
@@ -53,50 +52,7 @@ const cartSlice= createSlice({
   }
 })
 
-// create thunk
-export const sendCartData = (cart) => {
-  return async (dispatch) => {
-    dispatch(uiActions.showNotification({
-      status: 'pending',
-      title: 'Sending',
-      message: 'Sending cart data..'
-    }));
 
-    const sendRequest= async() => {
-      // called an API
-      const response= await fetch(
-        'https://react-movies-app-6b7ea-default-rtdb.firebaseio.com/cart-reduce.json', 
-        {
-          method: 'PUT',
-          body: JSON.stringify(cart)
-        }
-      );
-  
-      // check if response we havent got
-      if(!response.ok) {
-        throw new Error('Sending cart data failed.');
-      }
-    }
-
-    try {
-      await sendRequest();
-      
-      // dispatch success action
-      dispatch(uiActions.showNotification({
-        status: 'success',
-        title: 'Success!',
-        message: 'Sent cart data successfully!'
-      }));
-    } catch (error) {
-      // dispatch success action
-      dispatch(uiActions.showNotification({
-        status: 'error',
-        title: 'Error!',
-        message: 'Sending cart data failed!'
-      }));
-    }
-  }
-}
 
 export const cartActions= cartSlice.actions;
 export default cartSlice;
